requirements.txt
# Stock Analysis Project Dependencies
# Install with: pip install -r requirements.txt

alpaca-py>=0.21.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
scikit-learn>=1.3.0
plotly>=5.18.0
streamlit>=1.30.0
rich>=13.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0

